Moved Exchanger and LazyRegister to Edge module.
