+++ lib/concurrent-edge.rb
@@ -3,7 +3,7 @@ require 'concurrent'
 require 'concurrent/actor'
 require 'concurrent/agent'
 require 'concurrent/channel'
-require 'concurrent/exchanger'
-require 'concurrent/lazy_register'
 
+require 'concurrent/edge/exchanger'
 require 'concurrent/edge/future'
+require 'concurrent/edge/lazy_register'
