+++ spec/concurrent/edge/lazy_register_spec.rb
@@ -1,0 +1,10 @@
+module Concurrent
+  module Edge
+
+    describe LazyRegister do
+
+      pending
+
+    end
+  end
+end
