--- spec/concurrent/lazy_register_spec.rb
@@ -1,7 +1,0 @@
-module Concurrent
-  describe LazyRegister do
-
-    pending
-
-  end
-end
